Compute target/waypoint bearing on the typed map position
computeTargetBearing and the HUD waypoint-bearing producer rounded
flightMapPosition() to coarse map words before the delta, losing modern
fractional position ahead of the target-lock range gates and autopilot
guidance input. Both now run the delta through mapOffset()/mapRange() and
GuidanceMath::aimBearing, which is widened to double deltas: fixed narrows
to the computeBearing int endpoint (integer callers exact), modern keeps
the fraction through atan2.

Fixed is bit-exact: mapOffset sign-extends the int16 rep like the original
mapWordX/Y subtraction, fixed::computeBearing models the int16 entry
narrowing, and the g_targetRange store keeps its (int16) wrap.

Verification: fixed 59/59 (sortie parity covers target selection and the
autopilot HUD path), modern smoke, fractional aimBearing test, analyzer
clean on touched TUs.

+++ src/math/guidance.hpp
@@ -181,13 +181,15 @@ public:
 
     /* ---- Projectile guidance (updateThreatTargeting / fireMissile) ---- */
 
-    /* computeBearing endpoint: the aim bearing for a plain delta pair. Fixed
-     * runs the original LUT approximation; modern evaluates atan2 directly —
-     * the table + linear interpolation was the legacy limit. (0, 0) keeps the
-     * original's south answer. */
-    static Angle<B> aimBearing(int dx, int dy) {
+    /* computeBearing endpoint: the aim bearing for a delta pair. Fixed runs
+     * the original LUT approximation on the word-narrowed ints (integer
+     * callers are exact through the double parameter); modern evaluates
+     * atan2 directly — the table + linear interpolation was the legacy limit —
+     * and keeps fractional deltas. (0, 0) keeps the original's south answer. */
+    static Angle<B> aimBearing(double dx, double dy) {
         if constexpr (std::is_same_v<B, FixedBackend>)
-            return Angle<B>(fixed::computeBearing(dx, dy));
+            return Angle<B>(fixed::computeBearing(static_cast<int>(dx),
+                                                  static_cast<int>(dy)));
         else
             return Angle<B>(dx == 0 && dy == 0 ? 32768.0 * wordRadians : std::atan2(dx, dy));
     }
